+++ packages/blocks/core/CHANGELOG.md
@@ -43,6 +43,7 @@
 - `Post Excerpt` block: inner blocks improved.
 - `Post Terms` block: inner blocks improved.
 - `Post Title` block: inner blocks improved.
+- `Pullquote` block: inner blocks improved.
 
 
 ### Automated Tests